Update Google Analytics to merge Eberly & OTE adwords accounts.
Wouldn't it be nice if Google had a way to move properties & profiles from
one analytics account to another? Now we have to start with a fresh seminars
property in order to house it under the OTE account, which we renamed to Eberly,
because there was no way I was abandoning the blackboard analytics.
I should write a letter.