Add a path parameter to the default environment so rails:console and rails:dbconsole aren't dependent on the path on the host.
There isn't an elegant way to both have an interactive console session and
use execute (for its command parsing and sshkit command mappings;
understanding of an outer 'with', 'within', and 'on' blocks; and quiet
inclusion of the default environment parameters. So right now exec is
passed a long string that includes an explicit cd into $deploy/current and
setting of the path to the default_env[:path] value.