Fix Transcripts for New Users & Unify Transcript Layout for All Users
Fix the problem that if a user didn't exist and their transcript was
downloaded, it would throw an error.
Transcripts no longer distinguish between enrolled and unenrolled
participants it the FFP; if a user is not enrolled, the current FFP
requirement categories are displayed.
Finally update the "Worshop" category title to be "Workshops and
Other Programming", consistent with the web view; also actually
include any Other Programming items.